Moved exception to the project root directory
