+++ calendly/__init__.py
@@ -1,4 +1,5 @@
 from .utils.oauth2 import CalendlyOauth2
 from .calendly import CalendlyAPI
+from .exceptions import CalendlyException, CalendlyOauth2Exception
 
-__all__ = [CalendlyAPI, CalendlyOauth2]
+__all__ = [CalendlyAPI, CalendlyOauth2, CalendlyException, CalendlyOauth2Exception]
