+++ calendly/utils/api.py
@@ -1,5 +1,5 @@
 from typing import MutableMapping
-
+from calendly.exceptions import CalendlyOauth2Exception
 import requests
 
 __author__ = "laxmena <ConnectWith@laxmena.com>"
